% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcp.R
\name{createTable}
\alias{createTable}
\alias{dropTable}
\alias{checkTableExists}
\title{Create or drop table}
\usage{
createTable(connectargs, table, coltypes, ...)

dropTable(connectargs, table, ...)

checkTableExists(connectargs, table)
}
\arguments{
\item{connectargs}{named list of connection arguments. See \link[bcputility]{makeConnectArgs}.}

\item{table}{Name of the source table when importing from  SQL Server. For specifying the
schema in the table name see \code{<schema>.<table>} and if not specified the
default is "dbo".}

\item{coltypes}{character vector of data types with the column names as list/vector names.
Use \link[bcputility]{mapDataTypes} or refer to for proper format.}

\item{...}{arguments to pass to \link[base]{system2}}
}
\value{
No return value. Operations from bcp are printed to console; see
\code{...} to redirect output
}
\description{
Create or drop table
}
