% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcp.R
\name{SQLServerCLIVersions}
\alias{SQLServerCLIVersions}
\alias{bcpVersion}
\alias{sqlcmdVersion}
\title{Check bcp and sqlcmd versions}
\usage{
bcpVersion(...)

sqlcmdVersion(...)
}
\arguments{
\item{...}{arguments to pass \link[base]{system2}}
}
\description{
Check bcp and sqlcmd versions
}
